requirements.txt
# VisionAgent - Professional Multi-Modal AI Agent Framework

## Requirements

# Core Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6

# Computer Vision & AI
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.0.0
transformers>=4.35.0
face-recognition>=1.3.0

# Data Processing
pyyaml>=6.0
requests>=2.31.0
aiofiles>=23.2.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.6.0

# Optional Dependencies (for enhanced features)
# insightface>=0.7.3  # Advanced face recognition
# onnx>=1.15.0        # ONNX model support
# onnxruntime>=1.16.0 # ONNX runtime
# tensorrt>=8.6.0     # TensorRT acceleration (NVIDIA)

# Production Dependencies
gunicorn>=21.2.0
redis>=5.0.0         # For caching and job queues
celery>=5.3.0        # For background task processing
